Fix the paths to recorderWorker.js
This patch _also_ reverses part of the previous commit which had removed the
paths from `webpack.config.ts` and changed the paths in `src/index.html` to be
rooted under `shared` instead of `tutor`. With the paths added, the application
can find the missing javascript files and work as expected.